fix: `core_expressions` feature flag broken, move `overlay` into `core` functions (#15217)
* fix: remove core_expressions, move overlay to core

* fix license header

* fix header

* update doc

* Remove CI reference

---------
