+++ datafusion/functions/src/core/mod.rs
@@ -32,6 +32,7 @@ pub mod named_struct;
 pub mod nullif;
 pub mod nvl;
 pub mod nvl2;
+pub mod overlay;
 pub mod planner;
 pub mod r#struct;
 pub mod union_extract;
@@ -42,6 +43,7 @@ make_udf_function!(arrow_cast::ArrowCastFunc, arrow_cast);
 make_udf_function!(nullif::NullIfFunc, nullif);
 make_udf_function!(nvl::NVLFunc, nvl);
 make_udf_function!(nvl2::NVL2Func, nvl2);
+make_udf_function!(overlay::OverlayFunc, overlay);
 make_udf_function!(arrowtypeof::ArrowTypeOfFunc, arrow_typeof);
 make_udf_function!(r#struct::StructFunc, r#struct);
 make_udf_function!(named_struct::NamedStructFunc, named_struct);
@@ -71,6 +73,10 @@ pub mod expr_fn {
         nvl2,
         "Returns value2 if value1 is not NULL; otherwise, it returns value3.",
         arg1 arg2 arg3
+    ),(
+        overlay,
+        "replace the substring of string that starts at the start'th character and extends for count characters with new substring",
+        args,
     ),(
         arrow_typeof,
         "Returns the Arrow type of the input expression.",
@@ -115,6 +121,7 @@ pub fn functions() -> Vec<Arc<ScalarUDF>> {
         arrow_cast(),
         nvl(),
         nvl2(),
+        overlay(),
         arrow_typeof(),
         named_struct(),
         // Note: most users invoke `get_field` indirectly via field access
